trendpulse/requirements.txt
flask==3.0.3
flask-pymongo==2.3.0
pymongo==4.7.3
flask-bcrypt==1.0.1
flask-login==0.6.3
flask-session==0.8.0
vaderSentiment==3.3.2
python-dotenv==1.0.1
